bugfix: check for existance of class Bla000Peer
to allow the cli entrance to work when setting
up a SssSphocoaApp from scratch.

The idea is to make a plist with the vitals,
regenerate .SssStemplates files with those values.

Prepare tables in db.

run the app from cli (appdir/bin/run -interactive)
using the menu you should be able to reverse the
db, fixScheme as needed and create the proxy
classes (Propel).