introducing SssS_Cypher3
better encryption with built in tamper-test.
SssS_Cypher3 is an intermediary class that should not be
instantiated! Instead use either SssS_Cypher3openSSLcli
or SssS_Cypher3mcrypt. They can be used interchangeably
depending on what the system supports and how much data
needs to be de- or encrypted.

The naming is chosen in this fashion to allow a possible
third option: SssS_Cypher3openSSLphp which would use the
openssl-php-module